--- openwave/spacetime/__init__.py
@@ -1,7 +1,0 @@
-"""
-Energy Source modules for OpenWave simulations.
-
-This package contains the fundamental components of Spacetime simulation:
-- data_grids: Wave data structures & representations
-- wave_engines: Energy-Wave representations and calculations
-"""
